+++ src/main/java/cart/dao/member/MemberDao.java
@@ -1,0 +1,9 @@
+package cart.dao.member;
+
+import cart.domain.member.Member;
+import java.util.List;
+
+public interface MemberDao {
+
+    List<Member> findAll();
+}
