+++ src/main/java/cart/domain/member/Member.java
@@ -1,0 +1,26 @@
+package cart.domain.member;
+
+public class Member {
+
+    private final Long id;
+    private final Email email;
+    private final Password password;
+
+    public Member(final Long id, final Email email, final Password password) {
+        this.id = id;
+        this.email = email;
+        this.password = password;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public Email getEmail() {
+        return email;
+    }
+
+    public Password getPassword() {
+        return password;
+    }
+}
