--- src/main/java/cart/product/dao/MemberDao.java
@@ -1,9 +1,0 @@
-package cart.product.dao;
-
-import cart.product.domain.Member;
-import java.util.List;
-
-public interface MemberDao {
-
-    List<Member> findAll();
-}
